test(ios): move #4394 caret focus repro into caret.ts
Fold the caretFocus.ts reproduction spec back into the Caret describe
block in caret.ts and delete the standalone file. No behavior change; the
spec still asserts the keyboard opens on the pre-#4371 branch.
